% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadDates.R
\name{downloadDates}
\alias{downloadDates}
\title{Return download dates of included datasets}
\usage{
downloadDates()
}
\value{
For the Global Invasive Species Database, the Reptile Database and AmphibiaWeb, the date of 
download is returned, as these datasets are updated periodically. 
For the BirdLife Taxonomic Checklist, the Wilson & Reeder Mammals of the World, the IUCN geographic range 
datasets, the version or edition is returned.
}
\description{
Returns either the specific date that datasets were downloaded, or returns the dataset version.
}
\examples{
downloadDates()

}
\author{
Pascal Title
}
