% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData.R
\docType{data}
\name{rangeBuilder-example}
\alias{rangeBuilder-example}
\alias{crotalus}
\alias{gshhs}
\title{rangeBuilder datasets}
\description{
Included datasets in \code{rangeBuilder}
}
\details{
The \code{crotalus} dataset is the result of a query for genus Crotalus on the VertNet
	search portal (\url{http://portal.vertnet.org/search}), and has been thinned and lightly 
	filtered, to serve as an example dataset for this package.
	
	The \code{gshhs} dataset is a simplified version of the low resolution version of the 
	GSHHG (the Global Self-Consistent, Hierarchical, High-resolution Geography Database) 
	available for download from \url{https://www.soest.hawaii.edu/pwessel/gshhg/}.
}
\references{
Wessel, P., and W. H. F. Smith, A Global Self-consistent, Hierarchical, High-resolution 
	Shoreline Database, J. Geophys. Res., 101, 8741-8743, 1996.
}
