\name{rangeBuilder-package}
\alias{rangeBuilder-package}
\alias{rangeBuilder}
\docType{package}
\title{
rangeBuilder
}
\description{
Provides tools for filtering occurrence records, generating alpha-hull-derived range polygons and mapping species richness. 

}
\details{
\tabular{ll}{
Package: \tab rangeBuilder\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-01-02\cr
License: \tab GPL-2 | GPL-3\cr
}

See the \code{examples} section of \code{\link{richnessRaster}} for an example that uses most functions in this package.

}

\author{
Pascal Title <ptitle@umich.edu>
}
\references{
Davis Rabosky, A.R., C.L. Cox, D.L. Rabosky, P.O. Title, I.A. Holmes, A. Feldman and J.A. McGuire. \emph{In Review}. Coral snakes predict the evolution of mimicry across New World snakes. Nature Communications. 

}

\keyword{ package }
