\name{downloadDates}
\alias{downloadDates}
\title{
Return download dates of included datasets
}
\description{
Returns either the specific date that datasets were downloaded, or returns the dataset version.
}
\usage{
downloadDates()
}

\value{
For the Global Invasive Species Database, the Reptile Database and AmphibiaWeb, the date of download is returned, as these datasets are updated periodically. 
For the BirdLife Taxonomic Checklist and the Wilson & Reeder Mammals of the World, the version or edition is returned. 
}

\author{
	Pascal Title
}

\examples{
downloadDates()
}
 
