% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData.R
\docType{data}
\name{rangeBuilder-example}
\alias{rangeBuilder-example}
\alias{crotalus}
\title{rangeBuilder datasets}
\description{
Included datasets in \code{rangeBuilder}
}
\details{
The \code{crotalus} dataset is the result of a query for genus Crotalus on the VertNet
	search portal (\url{http://portal.vertnet.org/search}), and has been thinned and lightly 
	filtered, to serve as an example dataset for this package.
}
