\name{rangeBuilder-package}
\alias{rangeBuilder-package}
\alias{rangeBuilder}
\docType{package}
\title{
rangeBuilder
}
\description{
Provides tools for filtering occurrence records, standardizing countries and species names, generating alpha-hull-derived range polygons and mapping species distributions. 

}
\details{
\tabular{ll}{
Package: \tab rangeBuilder\cr
Type: \tab Package\cr
Version: \tab 1.4\cr
Date: \tab 2017-05-31\cr
License: \tab GPL-2 | GPL-3\cr
}


}

\author{
Pascal Title <ptitle@umich.edu>
}
\references{
Davis Rabosky, A.R., C.L. Cox, D.L. Rabosky, P.O. Title, I.A. Holmes, A. Feldman and J.A. McGuire. 2016. Coral snakes predict the evolution of mimicry across New World snakes. Nature Communications 7:11484. 

}

\keyword{ package }
