% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangeBuilder-package.R
\docType{package}
\name{rangeBuilder}
\alias{rangeBuilder}
\title{rangeBuilder}
\description{
Provides tools for filtering occurrence records, standardizing countries and 
species names, generating alpha-hull-derived range polygons and mapping species 
distributions.
}
\references{
Davis Rabosky, A.R., C.L. Cox, D.L. Rabosky, P.O. Title, I.A. Holmes, A. Feldman and J.A. 
McGuire. 2016. Coral snakes predict the evolution of mimicry across New World snakes. Nature 
Communications 7:11484.
}
\author{
Pascal Title <ptitle@umich.edu>
}
